fix: focus issue with nexted contenteditables on touch devices
Simple example: editor which uses contenteditable. Like Quill.
Quill creates a nested <p> element inside the div.ql-editor[contenteditable=true].
For some reasons `isContentEditable` for such a <p> will return `true`, thus making `isInput` to be true
in `onTouchEnd` handler. But the `activeElement` will be a top level drawer component since <p> is not focusable (?).

The solution is to check if the touched element is nested within the active element.